Add domain field to newly created assets
