+++ services/cargohold/src/CargoHold/App.hs
@@ -30,6 +30,7 @@ module CargoHold.App
     appLogger,
     requestId,
     settings,
+    localUnit,
 
     -- * App Monad
     AppT,
@@ -56,6 +57,7 @@ import Control.Monad.Trans.Resource (ResourceT, runResourceT, transResourceT)
 import Data.Default (def)
 import Data.Metrics.Middleware (Metrics)
 import qualified Data.Metrics.Middleware as Metrics
+import Data.Qualified
 import Imports hiding (log)
 import Network.HTTP.Client (ManagerSettings (..), requestHeaders, responseTimeoutMicro)
 import Network.HTTP.Client.OpenSSL
@@ -75,7 +77,8 @@ data Env = Env
     _appLogger :: Logger,
     _httpManager :: Manager,
     _requestId :: RequestId,
-    _settings :: Opt.Settings
+    _settings :: Opt.Settings,
+    _localUnit :: Local ()
   }
 
 makeLenses ''Env
@@ -86,7 +89,8 @@ newEnv o = do
   lgr <- Log.mkLogger (o ^. optLogLevel) (o ^. optLogNetStrings) (o ^. optLogFormat)
   mgr <- initHttpManager (o ^. optAws . awsS3Compatibility)
   ama <- initAws (o ^. optAws) lgr mgr
-  return $ Env ama met lgr mgr def (o ^. optSettings)
+  let loc = toLocalUnsafe (o ^. optSettings . Opt.setFederationDomain) ()
+  return $ Env ama met lgr mgr def (o ^. optSettings) loc
 
 initAws :: AWSOpts -> Logger -> Manager -> IO AWS.Env
 initAws o l m =
